backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Configuración y variables de entorno
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Ollama - Modelos de IA locales
ollama==0.1.7

# Utilidades
httpx>=0.25.2,<0.27.0
aiofiles==23.2.1
aiosqlite==0.19.0

# Validación y parsing
pydantic==2.5.3

# Desarrollo y testing (opcional)
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.2.0
ruff>=0.3.0,<0.4.0
